Move prompt/command row count into terminal command classes
